// src/app/models/patients/patient-hope.ts
export class PatientHope {
    patientId: number;
    patientOrganizationId: number;
    organizationId: number;
    name: string;
    mrNo: number;
    mrNoUnit: number;
    birthDate: string;
    address: string;
    mobileNo1: string;
    mobileNo2: string;
    patientStatusId: number;
    medicalRecordStatusId: number;
  }